requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
shap>=0.44.0
plotly>=5.18.0
pyyaml>=6.0
joblib>=1.3.0
rsconnect-python>=1.22.0
